Move Firestore metadata watch into DocumentMetadataStore; SortedDocuments delegates [CLUE-576]
getMSTSnapshotFromFBSnapshot, watchFirestoreMetaDataDocs, exemplarMetadataDocs,
firestoreMetadataDocs, metadataDocsFiltered, metadataDocsWithoutUnit, and docsReceived
are moved verbatim from SortedDocuments into DocumentMetadataStore. SortedDocuments
gains thin delegating getters/methods; ISortedDocumentsStores gains documentMetadata.
IDocumentMetadataStoreStores gains curriculumConfig and documents. inFlightPointReads
is marked non-observable. JSDoc updated to reflect post-move ownership.
Tests updated in sorted-documents.test.ts and document-group.test.ts to construct
a DocumentMetadataStore and wire it into mock stores.
