+++ src/models/stores/document-group.test.ts
@@ -10,6 +10,7 @@ import { GroupDocument, ProblemDocument } from "../document/document-types";
 import { ClassModel, ClassModelType, ClassUserModel } from "./class";
 import { GroupModel, GroupsModel, GroupsModelType, GroupUserModel } from "./groups";
 import { ISortedDocumentsStores, MetadataDocMapModel, SortedDocuments } from "./sorted-documents";
+import { DocumentMetadataStore } from "./document-metadata-store";
 import { DB } from "../../lib/db";
 import { Bookmark, Bookmarks } from "./bookmarks";
 
@@ -184,6 +185,7 @@ function addDocBookmarks(bookmarks: Bookmarks, bookmarkMap: Record<string, Array
 
 describe('DocumentGroup Model', () => {
   let sortedDocuments: SortedDocuments;
+  let documentMetadata: DocumentMetadataStore;
   let mockDocuments: DocumentModelType[];
   let mockGroups: GroupsModelType;
   let mockClass: ClassModelType;
@@ -196,6 +198,11 @@ describe('DocumentGroup Model', () => {
     const db = mock(DB);
     Object.setPrototypeOf(db, DB);
     bookmarks = new Bookmarks({db});
+
+    documentMetadata = new DocumentMetadataStore(
+      { db: {}, user: { classHash: "" }, documents: { exemplarDocuments: [] } } as any
+    );
+    documentMetadata.metadataDocsFiltered = MetadataDocMapModel.create(mockMetadataDocuments);
 
     const mockStores: DeepPartial<ISortedDocumentsStores> = {
       //DeepPartial allows us to not need to mock the "dB" and "appConfig" stores
@@ -210,11 +217,11 @@ describe('DocumentGroup Model', () => {
         customTagRecord: {},
         mergedWith: (configTags?: Record<string, string>) => ({ ...(configTags ?? {}) })
       },
-      bookmarks
+      bookmarks,
+      documentMetadata,
     };
 
     sortedDocuments = new SortedDocuments(mockStores as ISortedDocumentsStores);
-    sortedDocuments.metadataDocsFiltered = MetadataDocMapModel.create(mockMetadataDocuments);
   });
 
   describe("byBookMarked Function", () => {
@@ -523,7 +530,7 @@ describe('DocumentGroup Model', () => {
           problem: undefined
         }
       };
-      sortedDocuments.metadataDocsFiltered = MetadataDocMapModel.create(metadataWithNoProblem);
+      documentMetadata.metadataDocsFiltered = MetadataDocMapModel.create(metadataWithNoProblem);
 
       const byProblemDocs = sortedDocuments.sortBy("Problem");
       expect(byProblemDocs.length).toBe(4);
